fix location float problem + timezone bug

+++ backend/db/db_types/building_table_types.py
@@ -1,4 +1,4 @@
-from sqlalchemy import Column, Integer, String, DECIMAL, DateTime, Boolean
+from sqlalchemy import Column, Integer, String, DECIMAL, Float, DateTime, Boolean
 from sqlalchemy.ext.declarative import declarative_base
 
 
@@ -10,7 +10,7 @@ class Building(Base):
     building_id = Column(Integer, primary_key=True, autoincrement=True)
     building_name = Column(String, nullable=False)
     building_address = Column(String, nullable=False)
-    building_address_latitude = Column(DECIMAL(9,6), nullable=False)
-    building_address_longitude = Column(DECIMAL(9,6), nullable=False)
+    building_address_latitude = Column(Float, nullable=False)
+    building_address_longitude = Column(Float, nullable=False)
     manager_id = Column(Integer, nullable=False)
